add: WriteBoard의 복잡한 데이터 흐름을 쪼개기 위한 파일 분리
